requirements.txt
google-api-python-client==2.2.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
Jinja2==2.11.3
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.19.5
pandas==1.1.5
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
urllib3==1.26.4
xlrd==2.0.1
